Move vote counting logic into separate class
